+++ lib/src/tracker/udp_tracker_base.dart
@@ -1,5 +1,7 @@
 import 'dart:async';
 import 'dart:io';
+import 'dart:math';
+import 'dart:developer' as dev;
 
 import 'dart:typed_data';
 
@@ -37,7 +39,13 @@ mixin UDPTrackerBase {
   Uint8List _connectionId;
 
   /// 远程URL
-  Uri get uri;
+  // Uri get uri;
+
+  Future<List<CompactAddress>> get addresses;
+
+  bool _closed = false;
+
+  bool get isClosed => _closed;
 
   /// 获取当前transcation id，如果有就返回，表示当前通信还未完结。如果没有就重新生成
   List<int> get transcationId {
@@ -55,49 +63,83 @@ mixin UDPTrackerBase {
     return randomBytes(4);
   }
 
+  Timer _connectTimer;
+
+  int _connectRetryTimes = 0;
+
+  int maxConnectRetryTimes = 3;
+
+  Timer _announceTimer;
+
+  int _announceRetryTimes = 0;
+
+  final int _maxAnnounceRetryTimes = 2;
+
   /// 与Remote通讯的第一次连接
   ///
   /// Announce 和 Scrape通讯的时候，都必须要走这第一步，是固定的。
   ///
   /// 参数completer是一个`Completer`实例。用于截获发生的异常，并通过completeError截获
-  Future _connect(Map options) {
-    var uri = this.uri;
-    if (uri == null) throw ('目标地址Uri不能为空');
+  void _connect(
+      Map options, List<CompactAddress> address, Completer completer) {
+    if (isClosed) {
+      if (!completer.isCompleted) completer.completeError('Tracker closed');
+      return;
+    }
     var list = <int>[];
     list.addAll(START_CONNECTION_ID); //这是个magic id
     list.addAll(ACTION_CONNECT);
     list.addAll(transcationId);
     var messageBytes = Uint8List.fromList(list);
-    return _sendMessage(messageBytes, uri.host, uri.port);
+    try {
+      _sendMessage(messageBytes, address);
+      // 开始计时，超时重连
+      _connectTimer?.cancel();
+      _connectTimer =
+          Timer(Duration(seconds: 15 * pow(2, _connectRetryTimes)), () {
+        _connectRetryTimes++;
+        if (_connectRetryTimes >= maxConnectRetryTimes) {
+          if (!completer.isCompleted) {
+            completer.completeError('Retry too many times');
+          }
+          dev.log(
+              'connect 超时，超过最大限制:$_connectRetryTimes($maxConnectRetryTimes)，关闭',
+              name: runtimeType.toString());
+          close();
+        } else {
+          Timer.run(() => _connect(options, address, completer));
+        }
+      });
+      return;
+    } catch (e) {
+      if (!completer.isCompleted) completer.completeError(e);
+      close();
+    }
   }
 
   /// 和Remote通信的入口函数。返回一个Future
   Future<T> contactAnnouncer<T>(Map options) async {
+    if (isClosed) return null;
     var completer = Completer<T>();
+    var adds = await addresses;
+    if (adds == null || adds.isEmpty) {
+      close();
+      if (!completer.isCompleted) {
+        completer.completeError('InternetAddress cant be null');
+      }
+      return completer.future;
+    }
     _socket?.close();
     _socket = await RawDatagramSocket.bind(InternetAddress.anyIPv4, 0);
-
-    var eventStream = _socket.timeout(TIME_OUT, onTimeout: (e) {
-      if (!completer.isCompleted) {
-        close();
-        completer.completeError(e);
-      }
-    });
-    eventStream.listen((event) async {
+    _socket.listen((event) {
       if (event == RawSocketEvent.read) {
         var datagram = _socket.receive();
         if (datagram == null || datagram.data.length < 8) {
           close();
-          completer.completeError('Transaction ID from tracker is wrong');
+          completer.completeError('Wrong datas');
           return;
         }
-        try {
-          var re = await _processAnnounceResponseData(datagram.data, options);
-          if (re != null) completer.complete(re);
-        } catch (e) {
-          close();
-          completer.completeError(e);
-        }
+        _processAnnounceResponseData(datagram.data, options, adds, completer);
       }
     }, onError: (e) {
       close();
@@ -109,12 +151,7 @@ mixin UDPTrackerBase {
     });
 
     // 第一步，连接对方
-    try {
-      await _connect(options);
-    } catch (e) {
-      close();
-      completer.completeError(e);
-    }
+    _connect(options, adds, completer);
     return completer.future;
   }
 
@@ -124,7 +161,8 @@ mixin UDPTrackerBase {
 
   /// 处理一次通信最终从remote获得的数据.
   ///
-  dynamic processResponseData(Uint8List data, int action);
+  dynamic processResponseData(
+      Uint8List data, int action, Iterable<CompactAddress> addresses);
 
   ///
   /// 与announce和scrape通信的时候，在第一次连接成功后，第二次发送的数据是不同的。
@@ -133,16 +171,13 @@ mixin UDPTrackerBase {
 
   ///
   /// 第一次连接成功后，发送第二次信息
-  Future _announce(Uint8List connectionId, Map options) async {
+  void _announce(Uint8List connectionId, Map options,
+      List<CompactAddress> addresses) async {
     var message = generateSecondTouchMessage(connectionId, options);
-    var uri = this.uri;
-    if (uri == null) {
-      throw '目标地址Uri不能为空';
-    }
     if (message == null || message.isEmpty) {
       throw '发送数据不能为空';
     } else {
-      return _sendMessage(message, uri.host, uri.port);
+      _sendMessage(message, addresses);
     }
   }
 
@@ -150,7 +185,12 @@ mixin UDPTrackerBase {
   ///
   /// 该方法并不会直接去处理Remote返回的最终消息，而且固定了整个通信流程。
   /// 该方法会去处理在第一次发送信息后收到消息，然后到接收到第二次消息的整个过程
-  Future _processAnnounceResponseData(Uint8List data, Map options) async {
+  void _processAnnounceResponseData(Uint8List data, Map options,
+      List<CompactAddress> address, Completer completer) {
+    if (isClosed) {
+      if (!completer.isCompleted) completer.completeError('Tracker Closed');
+      return;
+    }
     var view = ByteData.view(data.buffer);
     var tid = view.getUint32(4);
     if (tid == transcationIdNum) {
@@ -158,34 +198,72 @@ mixin UDPTrackerBase {
       // 表明连接成功，可以进行announce
       if (action == 0) {
         _connectionId = data.sublist(8, 16); // 返回信息的第8-16位是下次连接的connection id
-        await _announce(_connectionId, options); // 继续，不要停
+        _connectTimer?.cancel();
+        _connectTimer = null;
+        _announce(_connectionId, options, address); // 继续，不要停
+        _announceTimer?.cancel();
+        _announceTimer =
+            Timer(Duration(seconds: 15 * pow(2, _announceRetryTimes)), () {
+          _announceRetryTimes++;
+          if (_announceRetryTimes >= _maxAnnounceRetryTimes) {
+            dev.log(
+                'announce 超时，超过最大限制 : $_announceRetryTimes($_maxAnnounceRetryTimes)，重新连接',
+                name: runtimeType.toString());
+            _announceRetryTimes = 0;
+            _announceTimer?.cancel();
+            _announceTimer = null;
+            Timer.run(() => _connect(options, address, completer));
+          } else {
+            Timer.run(() => _processAnnounceResponseData(
+                data, options, address, completer));
+          }
+        });
         return;
       }
       // 发生错误
       if (action == 3) {
+        var errorMsg = 'Unknown error';
+        try {
+          errorMsg = String.fromCharCodes(data.sublist(8));
+        } catch (e) {
+          //
+        }
+        if (!completer.isCompleted) {
+          completer.completeError(errorMsg);
+        }
         close();
-        var errorMsg = String.fromCharCodes(data.sublist(8));
-        throw errorMsg;
+        return;
       }
       // announce获得返回结果
-      close(); // Announce获得结果后就关闭socket不再监听。
-      return processResponseData(data, action);
+      var result = processResponseData(data, action, address);
+      completer.complete(result);
+      close();
+    } else {
+      if (!completer.isCompleted) {
+        completer.completeError('Transacation ID incorrect');
+      }
+      close();
     }
   }
 
-  /// 关闭套接字
+  /// 关闭连接以及清楚设置
   void close() {
+    _closed = true;
     _socket?.close();
     _socket = null;
+    _connectTimer?.cancel();
+    _connectTimer = null;
+    _announceTimer?.cancel();
+    _announceTimer = null;
+    _connectRetryTimes = 0;
+    _announceRetryTimes = 0;
   }
 
   /// 发送数据包到指定的ip地址
-  Future _sendMessage(Uint8List message, String host, int port) async {
-    // try {
-    var ips = await InternetAddress.lookup(host);
-    for (var i = 0; i < ips.length; i++) {
-      var ip = ips[i];
-      _socket?.send(message, ip, port);
-    }
+  void _sendMessage(Uint8List message, List<CompactAddress> addresses) {
+    if (isClosed) return;
+    addresses.forEach((element) {
+      _socket?.send(message, element.address, element.port);
+    });
   }
 }