+++ CHANGELOG.md
@@ -9,3 +9,9 @@
 ## 1.3.1
 
 - Change the tracker interfaces and add a new dependency
+
+## 1.3.4
+- Add IPv6 support
+- Make tracker and scrape can re-try
+- Change Readme and example
+- Delete useless files